drm/msm/sde: add support for qsync cmd mode
This change adds support to enable qsync feature for
command mode panels.
Feature allows command mode panels to update the refresh
rate to the slowest rate (i.e. 30fps) supported by the
panel whereas the transfer time still remains in the higher
frame rate supported by the panel (i.e. 60fps). Main goal
of this feature is prevent frame drops due the trigger
of the frame got delayed for couple of milliseconds in a
use case where the content of the screen it is being
updated at the highest frame rate.

Change-Id: I53a81b465d0ed8cdf1255561923f862403bb9ab6
Signed-off-by: Ingrid Gallardo <ingridg@codeaurora.org>
Signed-off-by: Steve Cohen <cohens@codeaurora.org>